restructured src folders to make better sense. commented out contact form for now
