+++ portfolio/src/main.jsx
@@ -12,12 +12,12 @@ import './fonts.css';
 
 //import page components for the routes to use
 import App from './App.jsx'
-import Home from './components/pages/home/Home.jsx'
-import Portfolio from './components/pages/portfolio/Portfolio.jsx'
-import About from './components/pages/about/About.jsx'
-import Contact from './components/pages/contact/Contact.jsx'
-import Resume from './components/pages/resume/Resume.jsx'
-import Error from './components/pages/error/Error.jsx'
+import Home from './pages/home/Home.jsx'
+import Portfolio from './pages/portfolio/Portfolio.jsx'
+import About from './pages/about/About.jsx'
+import Contact from './pages/contact/Contact.jsx'
+import Resume from './pages/resume/Resume.jsx'
+import Error from './pages/error/Error.jsx'
 
 //define the router itself
 const router = createBrowserRouter([
