+++ portfolio/src/pages/about/About.jsx
@@ -1,5 +1,5 @@
 import './About.css'
-import reactLogo from '../../../assets/react.svg'
+import reactLogo from '/react.svg'
 import viteLogo from '/vite.svg'
 
 export default function About() {
