+++ portfolio/src/pages/contact/Contact.css
@@ -1,5 +1,5 @@
 .contactBackground {
-    background-image: url('../../../assets/images/tall-bg-1.jpg');
+    background-image: url('../../assets/images/tall-bg-1.jpg');
     background-position:center;
     background-size:cover;
     padding:0px;
